+++ CHANGELOG.md
@@ -30,11 +30,12 @@ Thank you to all who have contributed!
 ### Added
 
 ### Changed
+- StaticTypeInferencer and PlanTyper will not raise an error when an expression is inferred to `NULL` or `unionOf(NULL, MISSING)`. In which cases the StaticTypeInferencer and PlanTyper will still raise the Problem Code `ExpressionAlwaysReturnsNullOrMissing` but the severity of the problem has been changed to warning. The In case of an expression always returns `MISSING`, problem code `ExpressionAlwaysReturnsMissing` will be raised. 
 
 ### Deprecated
 
 ### Fixed
-- Inferencer will not raise an error when an expression is inferred to `NULL` or `unionOf(NULL, MISSING)`
+
 ### Removed
 
 ### Security